Move Apple ID credential check/observation from `AppDelegate` to its own class so that it doesn't add a lot of code to `AppDelegate`.
