+++ WooCommerce/Classes/AppDelegate.swift
@@ -38,11 +38,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     ///
     private var storePickerCoordinator: StorePickerCoordinator?
 
-    /// Keychain access for SIWA auth token
+    /// Checks on whether the Apple ID credential is valid when the app is logged in and becomes active.
     ///
-    private lazy var keychain = Keychain(service: WooConstants.keychainServiceName)
-
-    private var cancellable: ObservationToken?
+    @available(iOS 13.0, *)
+    private lazy var appleIDCredentialChecker = AppleIDCredentialChecker()
 
     // MARK: - AppDelegate Methods
 
@@ -67,7 +66,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         setupKeyboardStateProvider()
         handleLaunchArguments()
         if #available(iOS 13.0, *) {
-            observeLoggedInStateForAppleIDObservations()
+            appleIDCredentialChecker.observeLoggedInStateForAppleIDObservations()
         }
 
         // Display the Authentication UI
@@ -140,7 +139,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         // This is done here so the check is done on app launch and app switching.
         if #available(iOS 13, *) {
-            checkAppleIDCredentialState()
+            appleIDCredentialChecker.checkAppleIDCredentialState()
         }
 
         RequirementsChecker.checkMinimumWooVersionForDefaultStore()
@@ -413,91 +412,5 @@ extension AppDelegate {
     func authenticatorWasDismissed() {
         setupPushNotificationsManagerIfPossible()
         RequirementsChecker.checkMinimumWooVersionForDefaultStore()
-    }
-}
-
-// MARK: - Authentication helpers
-//
-private extension AppDelegate {
-    func isLoggedIn() -> Bool {
-        ServiceLocator.stores.isAuthenticated
-    }
-
-    func logout() {
-        ServiceLocator.stores.deauthenticate()
-    }
-}
-
-// MARK: - Apple Account Handling
-//
-@available(iOS 13.0, *)
-private extension AppDelegate {
-    func observeLoggedInStateForAppleIDObservations() {
-        cancellable = ServiceLocator.stores.isLoggedIn.subscribe { [weak self] isLoggedIn in
-            if isLoggedIn {
-                self?.startObservingAppleIDCredentialRevoked()
-            } else {
-                self?.stopObservingAppleIDCredentialRevoked()
-            }
-        }
-    }
-
-    func checkAppleIDCredentialState() {
-        // If not logged in, remove the Apple User ID from the keychain, if it exists.
-        guard isLoggedIn() else {
-            removeAppleIDFromKeychain()
-            return
-        }
-
-        // Get the Apple User ID from the keychain
-        guard let appleUserID = keychain.wooAppleID else {
-            DDLogInfo("checkAppleIDCredentialState: No Apple ID found.")
-            return
-        }
-
-        // Get the Apple User ID state. If not authorized, log out the account.
-        WordPressAuthenticator.shared.getAppleIDCredentialState(for: appleUserID) { [weak self] (state, error) in
-            DDLogDebug("checkAppleIDCredentialState: Apple ID state: \(state.rawValue)")
-
-            switch state {
-            case .revoked:
-                DDLogInfo("checkAppleIDCredentialState: Revoked Apple ID. User signed out.")
-                self?.logOutRevokedAppleAccount()
-            default:
-                // An error exists only for the notFound state.
-                // notFound is a valid state when logging in with an Apple account for the first time.
-                if let error = error {
-                    DDLogDebug("checkAppleIDCredentialState: Apple ID state not found: \(error.localizedDescription)")
-                }
-                break
-            }
-        }
-    }
-
-    func startObservingAppleIDCredentialRevoked() {
-        WordPressAuthenticator.shared.startObservingAppleIDCredentialRevoked { [weak self] in
-            guard let self = self else {
-                return
-            }
-            if self.isLoggedIn() {
-                DDLogInfo("Apple credentialRevokedNotification received. User signed out.")
-                self.logOutRevokedAppleAccount()
-            }
-        }
-    }
-
-    func stopObservingAppleIDCredentialRevoked() {
-        WordPressAuthenticator.shared.stopObservingAppleIDCredentialRevoked()
-    }
-
-    func logOutRevokedAppleAccount() {
-        removeAppleIDFromKeychain()
-        DispatchQueue.main.async { [weak self] in
-            self?.logout()
-        }
-    }
-
-    func removeAppleIDFromKeychain() {
-        keychain.wooAppleID = nil
     }
 }
